Rename BaseJdbcPropertiesProvider to TypeHandlingJdbcPropertiesProvider
